Complete module user and product

+++ Library/src/main/java/com/build/core_restful/service/impl/PermissionServiceImpl.java
@@ -1,73 +1,71 @@
 package com.build.core_restful.service.impl;
 
 import com.build.core_restful.domain.Permission;
+import com.build.core_restful.domain.request.PermissionRequest;
+import com.build.core_restful.domain.response.PageResponse;
+import com.build.core_restful.domain.response.PermissionResponse;
 import com.build.core_restful.repository.PermissionRepository;
 import com.build.core_restful.service.PermissionService;
-import lombok.extern.slf4j.Slf4j;
+import com.build.core_restful.util.exception.NewException;
+import com.build.core_restful.util.mapper.PermissionMapper;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-
-@Slf4j
 @Service
 public class PermissionServiceImpl implements PermissionService {
-
     private final PermissionRepository permissionRepository;
+    private final PermissionMapper permissionMapper;
 
-    public PermissionServiceImpl(PermissionRepository permissionRepository) {
+    public PermissionServiceImpl(PermissionRepository permissionRepository, PermissionMapper permissionMapper) {
         this.permissionRepository = permissionRepository;
+        this.permissionMapper = permissionMapper;
     }
 
     @Override
-    public List<Permission> getAllPermissions() {
-        return permissionRepository.findAll();
+    public PageResponse<Object> getAllPermissions(Pageable pageable) {
+        Page<PermissionResponse> page = permissionRepository.findAll(pageable).map(permissionMapper::toPermissionResponse);
+        return PageResponse.builder()
+                .page(page.getNumber())
+                .size(page.getSize())
+                .content(page.getContent())
+                .build();
     }
 
     @Override
-    public boolean existById(Long id) {
-        return permissionRepository.existsById(id);
+    public PermissionResponse getPermissionById(Long id) {
+        return permissionRepository.findById(id)
+                .map(permissionMapper::toPermissionResponse)
+                .orElseThrow(() -> new NewException("Permission with id " + id + " not found"));
     }
 
     @Override
-    public boolean existByName(String name) {
-        return permissionRepository.existsByName(name);
+    public PermissionResponse createPermission(PermissionRequest request) {
+        if (permissionRepository.existsByName(request.getName())) {
+            throw new NewException("Permission with name " + request.getName() + " already exists!");
+        }
+        Permission permission = permissionMapper.toPermission(request);
+        return permissionMapper.toPermissionResponse(permissionRepository.save(permission));
     }
 
     @Override
-    public Permission getPermissionById(Long id) {
-        return permissionRepository.findById(id).orElse(null);
+    public PermissionResponse updatePermission(Long id, PermissionRequest request) {
+        Permission permission = permissionRepository.findById(id)
+                .orElseThrow(() -> new NewException("Permission with id " + id + " not found"));
+        permissionMapper.updatePermission(permission, request);
+        return permissionMapper.toPermissionResponse(permissionRepository.save(permission));
     }
 
     @Override
-    public Permission getPermissionByName(String name) {
-        return permissionRepository.findByName(name);
-    }
-
-    @Override
-    public Permission createPermission(Permission newPermission) {
-        return permissionRepository.save(newPermission);
-    }
-
-    @Override
-    public Permission updatePermission(Permission updatePermission) {
-        return permissionRepository.save(
-                Permission.builder()
-                        .id(updatePermission.getId())
-                        .name(updatePermission.getName())
-                        .apiPath(updatePermission.getApiPath())
-                        .method(updatePermission.getMethod())
-                        .module(updatePermission.getModule())
-                        .roles(updatePermission.getRoles())
-                        .build()
-        );
-    }
-
-    @Override
-    public void deletePermission(Long id) {
+    public boolean deletePermission(Long id) {
+        if (!permissionRepository.existsById(id)) {
+            throw new NewException("Permission with id " + id + " not found!");
+        }
         try {
             permissionRepository.deleteById(id);
-        } catch (Exception ex) {
-            log.error(ex.getMessage());
+            return true;
+        } catch (Exception e) {
+            return false;
         }
     }
 }
